+++ utils/Objects.js
@@ -126,7 +126,7 @@ var Objects;
      */
     function cloneDeep(source) {
         if (source == null) {
-            throw new TypeError("cloneDeep() source cannot be null");
+            return source;
         }
         var srcType;
         if (Array.isArray(source)) {
@@ -158,16 +158,21 @@ var Objects;
     Objects.cloneDeep = cloneDeep;
     function clone(source, srcKeys) {
         if (source == null) {
-            throw new TypeError("clone() source cannot be null");
+            return source;
         }
+        var srcType;
         if (Array.isArray(source)) {
             var res = [];
             Array.prototype.push.apply(res, source);
             return res;
         }
-        else {
+        else if ((srcType = Object.prototype.toString.call(source)) === "[object Date]") {
+            return new Date(source.getTime());
+        }
+        else if (srcType === "[object Object]") {
             return assign({}, source, srcKeys);
         }
+        return source;
     }
     Objects.clone = clone;
     function assign(target, source, srcKeys) {
@@ -211,25 +216,6 @@ var Objects;
         return target;
     }
     Objects.assignAll = assignAll;
-    /** Get a property from an object without the risk of an undefined error.
-     * Return null if either the object or the property are null or undefined.
-     * Example: {@code getProp(undefined, "alpha")}
-     * returns: {@code null}
-     * Or example: {@code getProp({ alpha: 342 }, "alpha")}
-     * returns: {@code 342}
-     *
-     * @param obj: the object to retrieve the property from
-     * @param propertyName: the name of the object property to retrieve
-     * @return the property retrieved from the object if both the object and property are not null, else null
-     */
-    function getProp(obj, propertyName) {
-        if (obj == null) {
-            return null;
-        }
-        var prop = obj[propertyName];
-        return prop == null ? null : prop;
-    }
-    Objects.getProp = getProp;
     /** Get multiple properties from an object without the risk of an undefined error.
      * Return an empty array if either the object or the list of property names are null or undefined.
      * Example: {@code getProps(undefined, ["alpha", "beta"])}